feat: attribute src
HTML attributes are special words used inside the opening tag of an element to control the element's behavior. The src attribute in an img element specifies the image's URL (where the image is located). An example of an img element using an src attribute: <img src="https://www.your-image-source.com/your-image.jpg">.

Add an src attribute to the existing img element that is set to the following URL: https://cdn.freecodecamp.org/curriculum/cat-photo-app/relaxing-cat.jpg.