feat: button-type attribute
Even though you added your button below the text input, they appear next to each other on the page. That's because both input and button elements are inline elements, which don't appear on new lines.

You learned previously that the button submits the form by default, but you can explicitly add the type attribute with the value submit to make it clearer. Go ahead and do this to specify where this button should submit the form.